requirements.txt
tensorflow==2.15.0
protobuf<4
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
pillow